As far as  i can tell, the DB and game setup works
As far as the DB and backend goes, everything is working as intended.  I need to still do the front end part and start to hook up everything as well as the CSS styling for it so its still not ready for next class. will probably be done monday But i got other things to do